+++ src/net/ftpworker.h
@@ -22,44 +22,49 @@
 
 #pragma once
 
+#include <cstddef>
+#include <cstdint>
+
 #include <circle/net/ipaddress.h>
 #include <circle/net/socket.h>
+#include <circle/netdevice.h>
 #include <circle/sched/task.h>
 #include <circle/string.h>
+
 #include "../config.h"
 #include "mdnspublisher.h"
 
 // TODO: These may be incomplete/inaccurate
 enum TFTPStatus
 {
-	FileStatusOk		= 150,
+	FileStatusOk = 150,
 
-	Success			= 200,
-	SystemType		= 215,
-	ReadyForNewUser		= 220,
-	ClosingControl		= 221,
-	TransferComplete	= 226,
-	EnteringPassiveMode	= 227,
-	UserLoggedIn		= 230,
-	FileActionOk		= 250,
-	PathCreated		= 257,
+	Success = 200,
+	SystemType = 215,
+	ReadyForNewUser = 220,
+	ClosingControl = 221,
+	TransferComplete = 226,
+	EnteringPassiveMode = 227,
+	UserLoggedIn = 230,
+	FileActionOk = 250,
+	PathCreated = 257,
 
-	PasswordRequired	= 331,
-	AccountRequired		= 332,
-	PendingFurtherInfo	= 350,
+	PasswordRequired = 331,
+	AccountRequired = 332,
+	PendingFurtherInfo = 350,
 
-	ServiceNotAvailable	= 421,
-	DataConnectionFailed	= 425,
-	FileActionNotTaken	= 450,
-	ActionAborted		= 451,
+	ServiceNotAvailable = 421,
+	DataConnectionFailed = 425,
+	FileActionNotTaken = 450,
+	ActionAborted = 451,
 
-	CommandUnrecognized	= 500,
-	SyntaxError		= 501,
-	CommandNotImplemented	= 502,
-	BadCommandSequence	= 503,
-	NotLoggedIn		= 530,
-	FileNotFound		= 550,
-	FileNameNotAllowed	= 553,
+	CommandUnrecognized = 500,
+	SyntaxError = 501,
+	CommandNotImplemented = 502,
+	BadCommandSequence = 503,
+	NotLoggedIn = 530,
+	FileNotFound = 550,
+	FileNameNotAllowed = 553,
 };
 
 enum class TTransferMode
@@ -80,60 +85,60 @@ struct TDirectoryListEntry;
 class CFTPWorker : protected CTask
 {
 public:
-	CFTPWorker(CSocket* pControlSocket, const char* pExpectedUser, const char* pExpectedPassword, CmDNSPublisher* pMDNSPublisher, CConfig* pConfig);
+	CFTPWorker(CSocket *pControlSocket, const char *pExpectedUser, const char *pExpectedPassword, CmDNSPublisher *pMDNSPublisher, CConfig *pConfig);
 	virtual ~CFTPWorker() override;
 
 	virtual void Run() override;
 
-	static u8 GetInstanceCount() { return s_nInstanceCount; }
+	static uint8_t GetInstanceCount() { return s_nInstanceCount; }
 
 private:
-	CSocket* OpenDataConnection();
+	CSocket *OpenDataConnection();
 
-	bool SendStatus(TFTPStatus StatusCode, const char* pMessage);
+	bool SendStatus(TFTPStatus StatusCode, const char *pMessage);
 
 	bool CheckLoggedIn();
 
 	// Directory navigation
-	CString RealPath(const char* pInBuffer) const;
-	const TDirectoryListEntry* BuildDirectoryList(size_t& nOutEntries) const;
+	CString RealPath(const char *pInBuffer) const;
+	const TDirectoryListEntry *BuildDirectoryList(size_t &nOutEntries) const;
 
 	// FTP command handlers
-	bool System(const char* pArgs);
-	bool Username(const char* pArgs);
-	bool Port(const char* pArgs);
-	bool Passive(const char* pArgs);
-	bool Password(const char* pArgs);
-	bool Type(const char* pArgs);
-	bool Retrieve(const char* pArgs);
-	bool Store(const char* pArgs);
-	bool Delete(const char* pArgs);
-	bool MakeDirectory(const char* pArgs);
-	bool ChangeWorkingDirectory(const char* pArgs);
-	bool ChangeToParentDirectory(const char* pArgs);
-	bool PrintWorkingDirectory(const char* pArgs);
-	bool List(const char* pArgs);
-	bool ListFileNames(const char* pArgs);
-	bool RenameFrom(const char* pArgs);
-	bool RenameTo(const char* pArgs);
-	bool Bye(const char* pArgs);
-	bool NoOp(const char* pArgs);
+	bool System(const char *pArgs);
+	bool Username(const char *pArgs);
+	bool Port(const char *pArgs);
+	bool Passive(const char *pArgs);
+	bool Password(const char *pArgs);
+	bool Type(const char *pArgs);
+	bool Retrieve(const char *pArgs);
+	bool Store(const char *pArgs);
+	bool Delete(const char *pArgs);
+	bool MakeDirectory(const char *pArgs);
+	bool ChangeWorkingDirectory(const char *pArgs);
+	bool ChangeToParentDirectory(const char *pArgs);
+	bool PrintWorkingDirectory(const char *pArgs);
+	bool List(const char *pArgs);
+	bool ListFileNames(const char *pArgs);
+	bool RenameFrom(const char *pArgs);
+	bool RenameTo(const char *pArgs);
+	bool Bye(const char *pArgs);
+	bool NoOp(const char *pArgs);
 
 	CString m_LogName;
 
 	// Authentication
-	const char* m_pExpectedUser;
-	const char* m_pExpectedPassword;
+	const char *m_pExpectedUser;
+	const char *m_pExpectedPassword;
 
 	// TCP sockets
-	CSocket* m_pControlSocket;
-	CSocket* m_pDataSocket;
-	u16 m_nDataSocketPort;
+	CSocket *m_pControlSocket;
+	CSocket *m_pDataSocket;
+	uint16_t m_nDataSocketPort;
 	CIPAddress m_DataSocketIPAddress;
 
 	// Command/data buffers
 	char m_CommandBuffer[FRAME_BUFFER_SIZE];
-	u8 m_DataBuffer[FRAME_BUFFER_SIZE];
+	uint8_t m_DataBuffer[FRAME_BUFFER_SIZE];
 
 	// Session state
 	CString m_User;
@@ -143,17 +148,17 @@ private:
 	CString m_CurrentPath;
 	CString m_RenameFrom;
 
-	CmDNSPublisher* m_pmDNSPublisher;
-	CConfig* m_pConfig;
+	CmDNSPublisher *m_pmDNSPublisher;
+	CConfig *m_pConfig;
 
-	static void FatFsPathToFTPPath(const char* pInBuffer, char* pOutBuffer, size_t nSize);
-	static void FTPPathToFatFsPath(const char* pInBuffer, char* pOutBuffer, size_t nSize);
+	static void FatFsPathToFTPPath(const char *pInBuffer, char *pOutBuffer, size_t nSize);
+	static void FTPPathToFatFsPath(const char *pInBuffer, char *pOutBuffer, size_t nSize);
 
-	static void FatFsParentPath(const char* pInBuffer, char* pOutBuffer, size_t nSize);
+	static void FatFsParentPath(const char *pInBuffer, char *pOutBuffer, int nSize);
 
-	static void FormatLastModifiedDate(u16 nDate, char* pOutBuffer, size_t nSize);
-	static void FormatLastModifiedTime(u16 nDate, char* pOutBuffer, size_t nSize);
+	static void FormatLastModifiedDate(uint16_t nDate, char *pOutBuffer, size_t nSize);
+	static void FormatLastModifiedTime(uint16_t nDate, char *pOutBuffer, size_t nSize);
 
 	static const TFTPCommand Commands[];
-	static u8 s_nInstanceCount;
+	static uint8_t s_nInstanceCount;
 };
